Integrate mpv
Made a build script that collects all dependencies, builds them for the specified architectures, and copies them to the libs folder of the Player project.
Made use of the preview cmake support to link the static libraries to a single shared lib.
Add beginning of code that interfaces with the native code.